Add Vue2-leaflet as UI components
Leaflet is a popular mapping (and not only) library, vue2-leaflet aims to allow the user to reason and configure leaflet trough the uses of components. As of version 2.x the library support tree-shaking to avoid to add unnecessary weight to the bundle.

vue2-leaflet has the full api documented, all of the component come with an example and a playground, dedicated documentation for nuxt and a FAQ section. a lot of examples are added with source code reachable in Git. Unit test are growing with dedicated contributors, issues are answered reasonably fast.

There is a good number of additional plugin for vue2-leaflet and are listed in our docs.

Note: I am not the author but I am the current maintainer, since the original author is MIA since almost one year.